Report error locations on errors.
It would appear that while Cryptol has entries for location information
in its AST, we are currently not generating these annotations.  So for
the time being we just keep track of the names of the declarations
we are working on, and report these on error, which is usually good
enough to figure out what's the problem.

See #6